+++ crates/core/src/config/custom.rs
@@ -13,6 +13,8 @@ const CUSTOM_CONFIG_RUNTIME_ISOLATE: &str = "runtime.isolate";
 pub const CUSTOM_CONFIG_EXTERNAL_RECORD: &str = "external.record";
 pub const CUSTOM_CONFIG_RESOLVE_DEDUPE: &str = "resolve.dedupe";
 pub const CUSTOM_CONFIG_CSS_MODULES_LOCAL_CONVERSION: &str = "css.modules.locals_conversion";
+pub const CUSTOM_CONFIG_PARTIAL_BUNDLING_GROUPS_ENFORCE_MAP: &str =
+  "partial_bundling.groups.enforce";
 
 pub fn get_config_runtime_isolate(context: &Arc<CompilationContext>) -> bool {
   if let Some(val) = context.config.custom.get(CUSTOM_CONFIG_RUNTIME_ISOLATE) {
@@ -28,8 +30,8 @@ pub fn get_config_external_record(config: &Config) -> ExternalConfig {
       return ExternalConfig::new();
     }
 
-    let external: HashMap<String, String> = serde_json::from_str(val)
-      .unwrap_or_else(|_| panic!("failed parse record external {val:?}"));
+    let external: HashMap<String, String> =
+      serde_json::from_str(val).unwrap_or_else(|_| panic!("failed parse record external {val:?}"));
 
     let mut external_config = ExternalConfig::new();