+++ js-plugins/less/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @farmfe/js-plugin-less
 
+## 2.0.0-nightly-20241022124925
+
+### Patch Changes
+
+- @farmfe/core@2.0.0-nightly-20241022124925
+
+## 2.0.0-nightly-20241022041556
+
+### Patch Changes
+
+- Updated dependencies [24347c9]
+  - @farmfe/core@2.0.0-nightly-20241022041556
+
+## 2.0.0-nightly-20241022014521
+
+### Patch Changes
+
+- Updated dependencies [a82a5b2]
+  - @farmfe/core@2.0.0-nightly-20241022014521
+
 ## 1.9.0
 
 ### Patch Changes